refactor: Clean up some.
